ejercicio caballeros

+++ UD2-BBDD/Ejercicio caballeros/index.php
@@ -14,6 +14,14 @@
     <div class="container mt-5">
         <h1 class="text-center mb-4">Lista de Caballeros</h1>
 
+        <!-- FILTRO -->
+        <form class="d-flex" role="search">
+            <input class="form-control me-2" type="search" placeholder="Search" aria-label="Search" name="buscar">
+            <button class="btn btn-outline-success" type="submit">Search</button>
+        </form>
+        <br>
+
+        <!-- TABLA -->
         <table class="table table-striped table-bordered">
             <thead class="bg-dark text-white">
                 <tr>
@@ -33,39 +41,70 @@
                     $resultado = $mysqli->query("SELECT * FROM caballeros");
 
                     while ($row = $resultado->fetch_assoc()) {
-                    ?>
-                </tr>
-                <tr>
-                    <td><?= $row["id"] ?></td>
-                    <td><?= $row["nombre"] ?></td>
-                    <td><?= $row["fuerza"] ?></td>
-                    <td><?php if ($row["activo"] == 1) {
-                            echo '<span class="badge bg-success">Sí</span>';
+                        if (isset($_GET['buscar'])) {
+                            $buscar = $_GET['buscar'];
+                            if (str_contains(strtolower($row["nombre"]), strtolower($buscar))) {
+                            ?>
+
+                            </tr>
+                            <tr>
+                                <td><?= $row["id"] ?></td>
+                                <td><?= $row["nombre"] ?></td>
+                                <td><?= $row["fuerza"] ?></td>
+                                <td><?php if ($row["activo"] == 1) {
+                                        echo '<span class="badge bg-success">Sí</span>';
+                                     } else {
+                                        echo '<span class="badge bg-danger">No</span>';
+                                     }
+                             ?> </td>
+                                <td>
+                                    <a href='verDatos_caballeros.php?id=<?= $row["id"] ?>' class="btn btn-info btn-sm">Ver</a>
+
+                                    <a href='update_form_caballeros.php?id=<?= $row["id"] ?>' class="btn btn-warning btn-sm">Modificar</a>
+
+                                    <a href='delete_caballeros.php?id=<?= $row["id"] ?>' class="btn btn-danger btn-sm">Eliminar</a>
+                                </td>
+
+                            </tr>
+                            <?php
+                            }
                         } else {
-                            echo '<span class="badge bg-danger">No</span>';
+                            ?>
+
+                            </tr>
+                            <tr>
+                                <td><?= $row["id"] ?></td>
+                                <td><?= $row["nombre"] ?></td>
+                                <td><?= $row["fuerza"] ?></td>
+                                <td><?php if ($row["activo"] == 1) {
+                                    echo '<span class="badge bg-success">Sí</span>';
+                                } else {
+                                    echo '<span class="badge bg-danger">No</span>';
+                                }
+                                ?></td>
+                                <td>
+                                    <a href='verDatos_caballeros.php?id=<?= $row["id"] ?>' class="btn btn-info btn-sm">Ver</a>
+
+                                    <a href='update_form_caballeros.php?id=<?= $row["id"] ?>' class="btn btn-warning btn-sm">Modificar</a>
+
+                                    <a href='delete_caballeros.php?id=<?= $row["id"] ?>' class="btn btn-danger btn-sm">Eliminar</a>
+                                </td>
+
+                            </tr>
+                            <?php
                         }
-                        ?></td>
-                    <td>
-                        <a href='verDatos_caballeros.php?id=<?= $row["id"] ?>' class="btn btn-info btn-sm">Ver</a>
+                    } ?>
 
-                        <a href='update_form_caballeros.php?id=<?= $row["id"] ?>' class="btn btn-warning btn-sm">Modificar</a>
-
-                        <a href='delete_caballeros.php?id=<?= $row["id"] ?>' class="btn btn-danger btn-sm">Eliminar</a>
-                    </td>
-
-                </tr>
-            <?php } ?>
-
-            <?php
-            $mysqli->close();
-            ?>
+    <?php
+    $mysqli->close();
+    ?>
 
             </tbody>
 
-            <tr>
-                <a href="form.html">Añadir caballero</a>
-            </tr>
         </table>
+        <tr>
+            <a href="form.html" class="btn btn-secondary btn-sm">Añadir caballero</a>
+        </tr>
     </div>
 
     <!-- Bootstrap JS -->
